+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Upgrading to 3.9.x versions requires that you upgrade to latest 3.5.x version first.
 
+- Migrate SPHINX_ constants to setup.php, #705
+
 [3.9.0]: https://github.com/eventum/eventum/compare/v3.8.4...master
 
 ## [3.8.4] - 2019-10-21